Rename commandbutton.js to command.js
